+++ ChangeLog.rst
@@ -50,7 +50,7 @@ Tik Manager Version History
 * **Fix.071:** Minor UI fixes
 * **Fix.072:** Houdini open file path fix ("\\" => "/")
 * **Fix.072:** Bug fix when canceling the item selection in project materials
-* **Fix.073:** Bug fix with default categories when trying to open scene manager from a non-scene manager project
+* **Fix.073:** Bug fix with default categories when trying to open pd manager from a non-pd manager project
 * **Fix.073:** Bug fix - Image Manager / querying shading groups
 * **Fix.074:** Bug fix - Image Manager / Unusable Referenced Render Layer checking error fixed
 * **Fix.075:** Bug fix - Houdini Module loading and importing modules setting $HIP location fixed